HS
The system has the facility to give a unique id for every patient and stores the details of
every patient and the staff automatically. It includes a search facility to know the current
status of each room. User can search availability of a doctor and the details of a patient
using the id.